Fixed TransparentGray 2.0
Not sure how much sense this makes though. `GrayA` seems to be the only
subtype of `TransparentGray`...
Also, I changed `ClipMinMax` to `ClampMinMax`, as it seems deprecated in
`Images.jl`.  But these lines mean that we need to depend on Images.jl,
which is fine. Especially if we have a lean package with only the image
types.